Move metadata package
