--- packages/kinvey-metadata/package.json
@@ -1,18 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-metadata",
-  "main": "src/index.js",
-  "dependencies": {
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-utils": "^3.9.3-beta.7",
-    "lodash": "4.17.4"
-  },
-  "devDependencies": {
-    "expect": "1.20.2",
-    "mocha": "4.0.1"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
